refactor: move `CompletionContext` from `Ast` to `shared`
