+++ main/src/ca/uwaterloo/flix/api/lsp/provider/completion/CompletionContext.scala
@@ -16,8 +16,6 @@
 package ca.uwaterloo.flix.api.lsp.provider.completion
 
 import ca.uwaterloo.flix.api.lsp.{Position, Range}
-import ca.uwaterloo.flix.language.CompilationMessage
-import ca.uwaterloo.flix.language.ast.Ast.SyntacticContext
 
 /**
   * Represents a completion context.
